+++ lib/api/routes/fetch.ts
@@ -20,6 +20,7 @@ const routeFetch = (zamza: Zamza) => {
                 "/api/fetch/:topic/paginate/stoe/:skipToIndex/:limit",
                 "/api/fetch/:topic/paginate/etos/:skipToIndex/:limit",
                 "/api/fetch/:topic/query/find",
+                "/api/fetch/:topic/query/count",
             ],
         });
     });
@@ -196,7 +197,7 @@ const routeFetch = (zamza: Zamza) => {
     router.post("/:topic/query/find", async (req, res) => {
 
         const {topic} = req.params;
-        const {query} = req.body;
+        const {query, limit = 512, skipToIndex = null, order = -1} = req.body;
 
         if (!res.locals.access.topicAccessAllowedForRequest(req, topic)) {
             res.status(403).json({
@@ -206,7 +207,28 @@ const routeFetch = (zamza: Zamza) => {
         }
 
         try {
-            res.status(200).json(await keyIndexModel.findForQuery(topic, query));
+            res.status(200).json(await keyIndexModel.findForQuery(topic, query, limit, skipToIndex, order, false));
+        } catch (error) {
+            res.status(500).json({
+                error: "An error occured " + error.message,
+            });
+        }
+    });
+
+    router.post("/:topic/query/count", async (req, res) => {
+
+        const {topic} = req.params;
+        const {query, limit = 512, skipToIndex = null, order = -1} = req.body;
+
+        if (!res.locals.access.topicAccessAllowedForRequest(req, topic)) {
+            res.status(403).json({
+                error: "Access not allowed, for this topic",
+            });
+            return;
+        }
+
+        try {
+            res.status(200).json(await keyIndexModel.findForQuery(topic, query, 0, null, 0, true));
         } catch (error) {
             res.status(500).json({
                 error: "An error occured " + error.message,