+++ lib/api/routes/index.ts
@@ -6,3 +6,4 @@ export * from "./produce";
 export * from "./manage";
 export * from "./hook";
 export * from "./replay";
+export * from "./state";