+++ lib/db/models/index.ts
@@ -4,3 +4,4 @@ export * from "./TopicMetadata";
 export * from "./LockModel";
 export * from "./HookModel";
 export * from "./ReplayModel";
+export * from "./StateModel";